seahub-extra/requirements.txt
python-cas
djangosaml2
cffi
